# docs/swagger.yaml
definitions:
  ncobase_common_resp.Exception:
    properties:
      code:
        description: Business code
        type: integer
      data:
        description: Response data
      errors:
        description: Validation errors
      message:
        description: Message
        type: string
      status:
        description: HTTP status
        type: integer
    type: object
  ncobase_common_types.JSON:
    additionalProperties: {}
    type: object
...
  structs.CreateTenantBody:
    properties:
      copyright:
        type: string
      created_by:
        type: string
      description:
        type: string
      disabled:
        type: boolean
      extras:
        $ref: '#/definitions/ncobase_common_types.JSON'
      keywords:
        items:
          type: string
        type: array
      logo:
        type: string
      logo_alt:
        type: string
      name:
        type: string
      order:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
...
  structs.ReadTenant:
    properties:
      copyright:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      disabled:
        type: boolean
      extras:
        $ref: '#/definitions/ncobase_common_types.JSON'
      id:
        type: string
      keywords:
        items:
          type: string
        type: array
      logo:
        type: string
      logo_alt:
        type: string
      name:
        type: string
      order:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
      user:
        $ref: '#/definitions/structs.User'
    type: object
...
  structs.UpdateTenantBody:
    properties:
      copyright:
        type: string
      created_by:
        type: string
      description:
        type: string
      disabled:
        type: boolean
      extras:
        $ref: '#/definitions/ncobase_common_types.JSON'
      id:
        type: string
      keywords:
        items:
          type: string
        type: array
      logo:
        type: string
      logo_alt:
        type: string
      name:
        type: string
      order:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
...
  structs.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      is_certified:
        type: boolean
      phone:
        type: string
      profile:
        $ref: '#/definitions/structs.UserProfile'
      updated_at:
        type: string
      username:
        type: string
    type: object
  structs.UserProfile:
    properties:
      about:
        type: string
      display_name:
        type: string
      id:
        type: string
      links:
        items:
          $ref: '#/definitions/ncobase_common_types.JSON'
        type: array
      short_bio:
        type: string
      thumbnail:
        type: string
    type: object
...
info:
  contact: {}
  description: Next-Gen Modular Framework
  title: Ncobase
  version: 0.1.0
paths:
...
  /v1/account/tenant:
    get:
      description: Retrieve the tenant associated with the current user.
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadTenant'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Get current user tenant
      tags:
      - account
...
  /v1/tenants:
    get:
      description: Retrieve a list of tenants.
      parameters:
      - in: query
        name: cursor
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/structs.ReadTenant'
            type: array
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: List tenants
      tags:
      - tenant
    post:
      consumes:
      - application/json
      description: Create a new tenant.
      parameters:
      - description: CreateTenantBody object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.CreateTenantBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadTenant'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Create tenant
      tags:
      - tenant
    put:
      consumes:
      - application/json
      description: Update the tenant information.
      parameters:
      - description: Tenant ID
        in: path
        name: slug
        required: true
        type: string
      - description: UpdateTenantBody object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateTenantBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadTenant'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Update tenant
      tags:
      - tenant
  /v1/tenants/{slug}:
    delete:
      description: Delete a specific tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Delete tenant
      tags:
      - tenant
    get:
      description: Retrieve information about a specific tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadTenant'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Get tenant
      tags:
      - tenant
  /v1/tenants/{slug}/assets:
    get:
      description: Retrieve a list of assets associated with a specific tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: List tenant assets
      tags:
      - tenant
  /v1/tenants/{slug}/groups:
    get:
      description: Retrieve a list of groups associated with a specific tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: List tenant groups
      tags:
      - tenant
  /v1/tenants/{slug}/menu:
    get:
      description: Retrieve the menu associated with a specific tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Get tenant menu
      tags:
      - tenant
  /v1/tenants/{slug}/setting:
    get:
      description: Retrieve the settings associated with a specific tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Get tenant setting
      tags:
      - tenant
  /v1/tenants/{slug}/users:
    get:
      description: Retrieve a list of users associated with a specific tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: List tenant users
      tags:
      - tenant
...
  /v1/users/{username}/tenant:
    get:
      description: Retrieve the tenant associated with the specified user.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadTenant'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Get user owned tenant
      tags:
      - user
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"